Addressed osx PR feedback.
